Fix UB in `SecStackSinglePageAlloc` when not using stds "magic" nightly allocator trait
The issue is that stds nightly `Allocator` is "magic" w.r.t. the `deallocate` function.
The pointer input variable (first input) received by the allocator, doesn't have the
provenance of the pointer that was passed to `deallocate`, but instead the potentially
larger provenance of the pointer that was returned by `allocate` for this allocation.

We round up allocation request sizes to multiples of 8, and then in `deallocate` we
zeroize this full (size multiple of 8) allocation. However, with a stable "clone"
of `Allocator` provided by the `allocator_api2` crate, there is no "magic", and the
pointer passed to `deallocate` can have a provenance to only the number of bytes that
were requested in tha `allocate` call, i.e. not rounded up to a multiple of 8.

Before the switch to `allocator_api2`, the test-suite didn't run on stable, so miri
missed this issue.