Introduce zeroizer based on (non-volatile) `write_bytes` and `asm!` optimisation barier and remove many slower ones
*Breaking:* Removes many zeroizers previously available.

This new zeroizer cannot be optimised out according to my current understanding of the op.sem. of Rust.
It can be implemented on stable, is target-independent, and faster than all our handwritten ones.
Therefore we can remove all the handwritten target-specific ones. This is good, since their implementation
turned out to be prone to mistakes.